2 Different simulations
Different information is provided in each simulation. 
The .xlsx file assumes that every produced unit should be used, otherwise it had no value at all. 
The .py simulation assumes that every produced unit (even though it is not required) has some market value, so it may be added to the revenues. 

Both simulations suggest that in this business model, whenever it is possible to ask for money and invest it in building upgrades, it may be done.