Simulation of the sales of Aerospace research.
It assumes that I am not paying for the development of the products used in the research. But for the development and upgrades of the Launchpads. Also this assumes that the most quality should be paid. Even though they are more expensive. 

The offered return over investmente through this model seems to be great when the launchpad increases its capabilities.